keep sptl/sptl consistent as uint8 with the rest of RTI2

+++ client_applications/converti2/converti2.h
@@ -51,8 +51,8 @@ class Converti2
 {
 private:
 	void setupSharedData();
-	uint32 sptl;
-	uint32 eptl;
+	uint8 sptl;
+	uint8 eptl;
 	bool showDistribution;
 	bool hasCheckPoints;
 	int argc;
